Crate syn_path

This crate contains macros to construct syn-types that contain paths inside a procedural macro.

The path! macro constructs a syn::Path.

The type_path! macro constructs a syn::TypePath.

The ty! macro constructs a syn::Type.

Macros

path

This macro takes paths of the form my_crate::my_mod::FooBar and ::my_crate::my_mod::FooBar and turns them into a syn::Path.

ty

This macro takes type paths of the form my_crate::my_mod::FooBar and <my_crate::my_mod::FooBar as my_crate::my_mod::MyTrait>::MyType and turns them into a syn::Type.

type_path

This macro takes type paths of the form my_crate::my_mod::FooBar and <my_crate::my_mod::FooBar as my_crate::my_mod::MyTrait>::MyType and turns them into a syn::TypePath.